+++ pymeasure/instruments/keithley/keithleyDMM6500.py
@@ -1172,7 +1172,7 @@ class KeithleyDMM6500(SCPIMixin, Instrument):
         and last channel in the list.
         Examples: ``1``, ``1,3,5``, ``1:2, 7:8``, or ``1:10``.
         """,
-        get_process=lambda x: x[-1].replace(")", ""),
+        get_process_list=lambda x: x[-1].replace(")", ""),
         separator="@",
     )